cleanup major: nlog => libnlog, everything is json
we needed messages to be an easier format because we need to be able to
go from a generic message into:

   something serializable over nanomsg / stdout and also something
   serializable to HTTP.

we do this now by using an alist and HTTP post is json body. this makes
everything more uniform and predictable. HTTP Headers are messy anyhow.

this causes no changes on the server-side because it just stores
everything anyway (headers, body (which is now in json)).

this also makes it possible to send multiple messages to the server in
one request: by issuing an array of message objects instead of a single
object, we know we are dealing with multiple objects. we may need to
restructure the nlogserver to deal with JSON and be less naive.